backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.13.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorclass==2.2.2
compressed-rtf==1.0.7
cryptography==46.0.5
distro==1.9.0
easygui==0.98.3
ebcdic==1.1.1
extract-msg==0.55.0
fastapi==0.129.0
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lark==1.3.1
lxml==6.0.2
msoffcrypto-tool==6.0.0
olefile==0.47
oletools==0.60.2
pcodedmp==1.2.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pypdf==6.7.0
python-docx==1.2.0
python-dotenv==1.2.1
python-magic==0.4.27
red-black-tree-mod==1.22
requests==2.32.5
rsa==4.9.1
RTFDE==0.1.2.2
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
tenacity==9.1.4
tuspyserver==4.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
